final-project/password_generator_test.py
import pytest
from tkinter import Tk
from tkinter import Label
from tkinter import StringVar
from tkinter import Entry
from unittest.mock import patch
from password_generator import populate_main_window, generate_password, clear

# Replace 'your_original_file_name' with the actual name of your original file

@pytest.fixture
def mock_tkinter(monkeypatch):
    # Mock tkinter.Tk to avoid opening a real window during testing
    def mock_tk():
        return Tk()
    
    monkeypatch.setattr("tkinter.Tk", mock_tk)

    # Mock tkinter.Label to create a dummy label
    def mock_label(master, **kwargs):
        return Label(master, **kwargs)
    
    monkeypatch.setattr("tkinter.Label", mock_label)

    # Mock tkinter.StringVar to create a dummy StringVar
    def mock_string_var(master, value=""):
        var = StringVar(master)
        var.set(value)
        return var
    
    monkeypatch.setattr("tkinter.StringVar", mock_string_var)

    # Mock tkinter.Entry to create a dummy entry widget
    def mock_entry(master, textvariable, **kwargs):
        return Entry(master, textvariable=textvariable, **kwargs)
    
    monkeypatch.setattr("tkinter.Entry", mock_entry)

def test_generate_password(mock_tkinter):
    # Mock user input to test generate_password function
    with patch("builtins.input", side_effect=["8", "y", "n", "y"]):
        # Call populate_main_window to set up the GUI elements
        frm_main = mock_tkinter
        populate_main_window(frm_main)

        # Call generate_password to test password generation
        generate_password(None)  # Pass None as event for testing

        # Get the generated password from the label
        password_label = frm_main.children['!lbl_password_gen']  # Replace 'label2' with the actual label name
        generated_password = password_label.cget("text")

        # Assert statements to test the generated password
        assert isinstance(generated_password, str)
        assert len(generated_password) == 8

def test_clear(mock_tkinter):
    # Mock user input to test clear function
    with patch("builtins.input", side_effect=["8", "y", "n", "y"]):
        # Call populate_main_window to set up the GUI elements
        frm_main = mock_tkinter
        populate_main_window(frm_main)

        # Call clear function to test clearing inputs and outputs
        clear()

        # Get the password label text after clearing
        password_label = frm_main.children['!lbl_password_gen']  # Replace 'label2' with the actual label name
        cleared_password = password_label.cget("text")

        # Assert statements to test clearing
        assert cleared_password == ""
        assert frm_main.children['!ent_password_length'].get() == ""  # Assuming entry widget ID is '!entry'

# Add more test cases for other functions as needed
